chore: setting for navigation
