+++ mainPage/animation/leftCloud.tsx
@@ -1,28 +1,28 @@
 import React, { useRef, useEffect } from "react";
 import { Image, Animated, Dimensions, Easing } from "react-native";
 
-type RightCloudProps = {
+type LeftCloudProps = {
   topValue: number;
   durationValue: number;
   delayValue: number;
 };
 
-export default function RightCloud({
+export default function LeftCloud({
   topValue,
   durationValue,
   delayValue,
-}: RightCloudProps) {
+}: LeftCloudProps) {
   let windowWidth = Dimensions.get("window").width;
-  const ref = useRef(new Animated.Value(windowWidth)).current;
+  const ref = useRef(new Animated.Value(0)).current;
 
   useEffect(() => {
     Animated.loop(
       Animated.timing(ref, {
-        toValue: -110,
+        toValue: windowWidth + 110,
         duration: durationValue,
         delay: delayValue,
-        useNativeDriver: false,
-        easing: Easing.linear,
+        useNativeDriver: false, // true시 loop가 안됨
+        easing: Easing.linear, // 애니메이션 일정 속도
       })
     ).start();
   }, [ref]);
@@ -30,6 +30,7 @@ export default function RightCloud({
   const styles: object = {
     position: "absolute",
     top: topValue,
+    left: -110,
     width: 104,
     height: 64,
     transform: [{ translateX: ref }],
@@ -38,8 +39,12 @@ export default function RightCloud({
   return (
     <Animated.View style={styles}>
       <Image
-        style={{ height: "100%", width: "100%", resizeMode: "contain" }}
-        source={require("../images/cloudReverse.png")}
+        style={{
+          height: "100%",
+          width: "100%",
+          resizeMode: "contain",
+        }}
+        source={require("../../images/cloud.png")}
       />
     </Animated.View>
   );
