+++ mainPage/animation/rightCloud.tsx
@@ -1,28 +1,28 @@
 import React, { useRef, useEffect } from "react";
 import { Image, Animated, Dimensions, Easing } from "react-native";
 
-type LeftCloudProps = {
+type RightCloudProps = {
   topValue: number;
   durationValue: number;
   delayValue: number;
 };
 
-export default function LeftCloud({
+export default function RightCloud({
   topValue,
   durationValue,
   delayValue,
-}: LeftCloudProps) {
+}: RightCloudProps) {
   let windowWidth = Dimensions.get("window").width;
-  const ref = useRef(new Animated.Value(0)).current;
+  const ref = useRef(new Animated.Value(windowWidth)).current;
 
   useEffect(() => {
     Animated.loop(
       Animated.timing(ref, {
-        toValue: windowWidth + 110,
+        toValue: -110,
         duration: durationValue,
         delay: delayValue,
-        useNativeDriver: false, // true시 loop가 안됨
-        easing: Easing.linear, // 애니메이션 일정 속도
+        useNativeDriver: false,
+        easing: Easing.linear,
       })
     ).start();
   }, [ref]);
@@ -30,7 +30,6 @@ export default function LeftCloud({
   const styles: object = {
     position: "absolute",
     top: topValue,
-    left: -110,
     width: 104,
     height: 64,
     transform: [{ translateX: ref }],
@@ -39,12 +38,8 @@ export default function LeftCloud({
   return (
     <Animated.View style={styles}>
       <Image
-        style={{
-          height: "100%",
-          width: "100%",
-          resizeMode: "contain",
-        }}
-        source={require("../images/cloud.png")}
+        style={{ height: "100%", width: "100%", resizeMode: "contain" }}
+        source={require("../../images/cloudReverse.png")}
       />
     </Animated.View>
   );
