+++ mainPage/footer/mapButton.tsx
