+++ mainPage/main/clothes/clothesItem.tsx
